chore: add ruff, and fix everything it found
Nothing linted this tree before. `ruff` covers pyflakes, pycodestyle, isort,
bugbear, pyupgrade and comprehensions in one tool; it runs in CI ahead of the
suite, because a lint failure is a one-line fix and finding it after a
four-minute test run wastes the time it was meant to save. `publish` now waits
on it as well as on the tests.

**`target-version = "py38"`, which is deliberately not `requires-python`.**
`muf` needs 3.10; `services/agent` is copied to the acquisition laptop and run
by the `.venv38` there. With a 3.10 target ruff wanted to rewrite
`typing.Callable` to `collections.abc.Callable` -- not subscriptable before
3.9 -- and to add `zip(..., strict=)`, which is 3.10. Both would have passed
CI on 3.12 and failed on the station, where a traceback is an AnyDesk session
and an outage rather than a red tick.
`tests/test_agent.py::test_the_agent_stays_within_python_38` parses every
agent module at 3.8, so the same breakage cannot arrive by any other route.

Four rules are ignored, each with its reason in pyproject.toml: the
ambiguous-unicode checks (the Cyrillic in `ru.py` is the product), `B008`
(`Depends()` in a default argument is how FastAPI declares a dependency, so
flagging it flags every route), `SIM105`, and `UP031` (its fix emits
`.format()` rather than an f-string and mangles the wrapped multi-line
messages this tree writes).

The other 352 findings are fixed. The ones that were not cosmetic:

* `services/agent/runner.py` imported `Path` and no longer used it, and
  `health.py` imported `control` inside `band()` without using it;
* `tests/test_cli.py` asserted `pytest.raises(Exception)` twice, which passes
  on any failure at all -- it names `argparse.ArgumentTypeError` now;
* `tests/test_loader.py` matched `"no reader for .dat"`, where the unescaped
  `.` matches any character;
* four dead locals, two exceptions re-raised inside an `except` without
  `from`, `l` as a variable name in three places, and 72 over-length lines.

Line endings are untouched: seven files in this tree are CRLF, and an earlier
pass through them normalised the lot to LF, which is a mass change with no
business hiding inside a lint commit.

+++ muf/calibrate.py
@@ -260,8 +260,8 @@ def gate_indices(
             f"(axis spans {axis_min:.0f}..{axis_max:.0f} km)"
         )
 
-    i_lo = max(0, min(int(math.ceil((half_span - hi_km) / step)), n_range - 1))
-    i_hi = max(0, min(int(math.floor((half_span - lo_km) / step)), n_range - 1))
+    i_lo = max(0, min(math.ceil((half_span - hi_km) / step), n_range - 1))
+    i_hi = max(0, min(math.floor((half_span - lo_km) / step), n_range - 1))
     if i_hi < i_lo:
         raise ValueError(
             f"empty range gate {lo_km:.0f}..{hi_km:.0f} km "
